move shared fns to utils module
