+++ relay-server/src/metrics_extraction/transactions/mod.rs
@@ -3,7 +3,7 @@ use std::collections::BTreeMap;
 use relay_common::{DurationUnit, EventType, SpanStatus, UnixTimestamp};
 use relay_dynamic_config::{AcceptTransactionNames, TaggingRule, TransactionMetricsConfig};
 use relay_general::protocol::{
-    AsPair, Context, ContextInner, Event, Span, TraceContext, TransactionSource,
+    AsPair, Context, ContextInner, Event, TraceContext, TransactionSource,
 };
 use relay_general::store;
 use relay_general::types::Annotated;
@@ -16,7 +16,7 @@ use crate::metrics_extraction::transactions::types::{
     TransactionMeasurementTags, TransactionMetric,
 };
 use crate::metrics_extraction::utils::{
-    extract_transaction_op, get_eventuser_tag, get_trace_context,
+    extract_http_status_code, extract_transaction_op, get_eventuser_tag, get_trace_context,
 };
 use crate::metrics_extraction::IntoMetric;
 use crate::statsd::RelayCounters;
@@ -65,82 +65,6 @@ fn extract_geo_country_code(event: &Event) -> Option<String> {
     if let Some(user) = event.user.value() {
         if let Some(geo) = user.geo.value() {
             return geo.country_code.value().cloned();
-        }
-    }
-
-    None
-}
-
-/// Extract the HTTP status code from the span data.
-pub(crate) fn http_status_code_from_span(span: &Span) -> Option<String> {
-    // For SDKs which put the HTTP status code into the span data.
-    if let Some(status_code) = span
-        .data
-        .value()
-        .and_then(|v| {
-            v.get("http.response.status_code")
-                .or_else(|| v.get("status_code"))
-        })
-        .and_then(|v| v.as_str())
-        .map(|v| v.to_string())
-    {
-        return Some(status_code);
-    }
-
-    // For SDKs which put the HTTP status code into the span tags.
-    if let Some(status_code) = span
-        .tags
-        .value()
-        .and_then(|tags| tags.get("http.status_code"))
-        .and_then(|v| v.as_str())
-        .map(|v| v.to_owned())
-    {
-        return Some(status_code);
-    }
-
-    None
-}
-
-/// Extracts the HTTP status code.
-pub(crate) fn extract_http_status_code(event: &Event) -> Option<String> {
-    if let Some(spans) = event.spans.value() {
-        for span in spans {
-            if let Some(span_value) = span.value() {
-                if let Some(status_code) = http_status_code_from_span(span_value) {
-                    return Some(status_code);
-                }
-            }
-        }
-    }
-
-    // For SDKs which put the HTTP status code into the breadcrumbs data.
-    if let Some(breadcrumbs) = event.breadcrumbs.value() {
-        if let Some(values) = breadcrumbs.values.value() {
-            for breadcrumb in values {
-                // We need only the `http` type.
-                if let Some(crumb) = breadcrumb
-                    .value()
-                    .filter(|bc| bc.ty.as_str() == Some("http"))
-                {
-                    // Try to get the status code om the map.
-                    if let Some(status_code) = crumb.data.value().and_then(|v| v.get("status_code"))
-                    {
-                        return status_code.value().and_then(|v| v.as_str()).map(Into::into);
-                    }
-                }
-            }
-        }
-    }
-
-    // For SDKs which put the HTTP status code in the `Response` context.
-    if let Some(contexts) = event.contexts.value() {
-        let response = contexts.get("response").and_then(Annotated::value);
-        if let Some(ContextInner(Context::Response(response_context))) = response {
-            let status_code = response_context
-                .status_code
-                .value()
-                .map(|code| code.to_string());
-            return status_code;
         }
     }
